feat: validation, before delete
유효성 검사 추가-
에러핸들링 추가-
vote delete fetch구현 직전
