+++ controller/signupController.js
@@ -1,0 +1,21 @@
+exports.showSignup = (req, res, next) => {
+  res.render("signup", {
+    message: null,
+  });
+};
+
+exports.createUser = async (req, res, next) => {
+  try {
+    const { email, password } = req.body;
+
+    await User.create({
+      email,
+      password,
+    });
+
+    res.redirect("/login");
+
+  } catch (error) {
+    next(error);
+  }
+};
