+++ package.json
@@ -17,6 +17,7 @@
     "express": "~4.16.1",
     "express-session": "^1.17.2",
     "express-validator": "^6.14.0",
+    "joi": "^17.5.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.1.5"
   },
