+++ public/javascripts/deleteVote.js
@@ -1,0 +1,7 @@
+const deleteButton = document.querySelector(".delete");
+deleteButton.addEventListener("click", () => {
+  fetch("/movies")
+  .then(function(response) {
+    return response.json();
+  })
+});
