+++ views/error.ejs
@@ -1,3 +1,4 @@
+<%- include("header.ejs") %>
 <h1><%= message %></h1>
 <h2><%= error.status %></h2>
 <pre><%= error.stack %></pre>
