+++ views/newVote.ejs
@@ -21,6 +21,6 @@
     <% if (message) { %>
       <p><%= message %></p>
     <% } %>
-    <script type="text/javascript" src="/javascripts/newVote.js"></script>
+    <script src="/javascripts/newVote.js"></script>
   </body>
 </html>
