+++ views/selectedVote.ejs
@@ -10,7 +10,7 @@
     <% if (isCreator) { %>
       <h2>투표 작성자 입니다.</h2>
       <form method="post" action="">
-        <input type="submit" value="delete" name="action">
+        <input type="submit" class="delete" value="delete" name="action">
       </form>
     <% } %>
     <% if (isCreator || isExpired) { %>
